[DEV-1.4] Fix bug preventing speakwrite's response from being stored in the db
