+++ src/components/HomePage.js
@@ -2,46 +2,12 @@
 
 import TextBlock from '@/components/TextBlock';
 import FilePanel from '@/components/FilePanel';
-import { useState, useEffect } from 'react';
-import { createUser, createSession } from '@/lib/api';
+import { useState } from 'react';
 
 export default function HomePage() {
   
   const [isOpen, setIsOpen] = useState(true);
   const [fileTitle, setFileTitle] = useState("Session 1");
-  const [c_uid, setCuid] = useState(null);
-  const [c_sid, setCsid] = useState(null);
-
-
-
-  useEffect(() => {
-    async function intializeUser() {
-      const user = await createUser("John Doe", "a@b.c", "12345678");
-    
-
-      if (user && user.id){
-        setCuid(user.id)
-        console.log(user)
-      }
-   
-    }
-    intializeUser();
-  }, []);
- 
-  useEffect(() => {
-    async function intializeSess() {
-
-        if (c_uid != null){
-        const session = await createSession({user_id:c_uid, context:{}});
-        
-        if (session && session.session_id){
-          setCsid(session.session_id)
-          console.log(session)
-        }
-    } }
-    
-    intializeSess();
-  }, [c_uid]);
 
 
   function togglePanel(){
@@ -58,7 +24,7 @@ export default function HomePage() {
           </button>
         )}
         {isOpen && <FilePanel fileTitle={fileTitle} onClose={togglePanel} />}
-        <TextBlock setFileTitle={setFileTitle} c_sid={c_sid}/>
+        <TextBlock setFileTitle={setFileTitle}/>
     </div>
   );
 }
